fix(logging): annotate the printf-like entry points, and fix what the checking found (#216)
dbg_printf, __syslog and the printf/sprintf/snprintf/fprintf/v*
family now carry __attribute__((format(printf, ...))), so every
format/argument mismatch in the tree is a compile-time error under
the existing -Werror.

The annotation found four real bugs, live at runtime:

- vfs_mkdir/rmdir/creat/stat/symlink logged the superblock failure
  with a %s and no argument, printing stack contents as the path.
- ftok's stat failure printed %s with no argument, same class.
- vfs_readlink logged a size_t bufsize through %d.
- shell's SIGCHLD sigaction failure logged the signal number
  through %s, and its .shellrc message had a conversion too many.

The rest of the ~120 diagnostics are type drift between specifier
and argument: size_t/ssize_t/long now printed with %lu/%ld, int-sized
values with %d/%u, typed pointers passed to %p cast to (void *)
(through an integer for the function pointers, which -Wpedantic
refuses to convert directly), and PRINT_REGS's %-0N flags cleaned of
the '0' that '-' already ignored.

Verified with GCC 13 and Clang 18 under the project's warning set.

Fixes #216.

+++ kernel/src/elf/elf.c
@@ -353,14 +353,14 @@ int elf_load_file(mm_struct_t *mm, vfs_file_t *file, uint32_t *entry)
         pr_err(
             "Failed to allocate %d bytes of memory for reading the file "
             "`%s`.\n",
-            stat_buf.st_size, file->name);
+            (int)stat_buf.st_size, file->name);
         return -ENOMEM;
     }
     // Clean the memory.
     memset(buffer, 0, stat_buf.st_size);
     // Read the file.
     if (vfs_read(file, buffer, 0, stat_buf.st_size) != stat_buf.st_size) {
-        pr_err("Failed to read %d bytes from the file `%s`.\n", stat_buf.st_size, file->name);
+        pr_err("Failed to read %d bytes from the file `%s`.\n", (int)stat_buf.st_size, file->name);
         goto return_error_free_buffer;
     }
     // The first thing inside the file is the ELF header.
